move ColorPicker component to own folder
