+++ src/components/ColorPicker/ColorPicker.scss
@@ -1,4 +1,4 @@
-@import "src/constants/style";
+@import "../../constants/style";
 
 .column__header-color-picker-dropdown {
   z-index: 220;
